requirement.txt
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
pillow==11.1.0
